wip trying to fix quiz index white screening
