+++ src/Kambaz/Courses/Quizzes/index.tsx
@@ -23,7 +23,7 @@ export default function Quizzes() {
     };
     useEffect(() => {
         fetchQuizzes();
-    }, []);
+    }, [quizzes]);
 
     return (
         <div>
